requirements.txt
modal>=0.66.0
supabase>=2.0.0
httpx>=0.27.0
stripe>=8.0.0
cryptography>=42.0.0
python-dotenv>=1.0.0
resend>=2.0.0
